requirements.txt
google-api-python-client
google_auth_oauthlib
pytelegrambotapi
celery
redis
python-decouple